// source/frontend/src/models/api/generated/ApiGen_CodeTypes_FileNumberTypes.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
export enum ApiGen_CodeTypes_FileNumberTypes {
  LISNO = 'LISNO',
  OTHER = 'OTHER',
  PROPNEG = 'PROPNEG',
  PSNO = 'PSNO',
  PUBWORKS = 'PUBWORKS',
  REGLSLIC = 'REGLSLIC',
  RESERVE = 'RESERVE',
}